Fix metadata by separating root layout into client and server
